requirements.txt
Django==1.8.2
django-crispy-forms==1.4.0
wsgiref==0.1.2
